+++ HISTORY.rst
@@ -2,6 +2,11 @@
 History
 =======
 
+2.3.2 (2022-01-28)
+------------------
+* The run() function now understands stdin=subprocess.DEVNULL to close the subprocess stdin,
+  rather than to connect through the existing stdin, which is the current default
+
 2.3.1 (2021-10-25)
 ------------------
 * Add Python 3.10 support